Internal function

Communicates between cursus and omnibus by coordinating a chromosome-wide analysis.

intern.chromo(Y, Ystart, Yend, X, Xloc,
             window, offset, group, perm,
             nodes, phi, kind)

Arguments

Y
RNA-Seq data: numeric matrix with q rows (genes) and n columns (samples); or a SummarizedExperiment object

Ystart
start location of genes: numeric vector of length q

Yend
end location of genes: NULL or numeric vector of length q

X
genomic profile: numeric matrix with p rows (covariates) and n columns (samples)

Xloc
location covariates: numeric vector of length p

window
maximum distance: non-negative real number

offset
numeric vector of length n

group
confounding variable: factor of length n

perm
number of iterations: positive integer

nodes
number of cluster nodes for parallel computation

phi
dispersion parameters: vector of length q

kind
computation : number between 0 and 1

Value

The function returns a dataframe, with the p-value in the first column, and the test statistic in the second column.

Examples

# simulate high-dimensional data
n <- 30
q <- 10
p <- 100
set.seed(1)
Y <- matrix(rnbinom(q*n,mu=10,
    size=1/0.25),nrow=q,ncol=n)
X <- matrix(rnorm(p*n),nrow=p,ncol=n)
Yloc <- seq(0,1,length.out=q)
Xloc <- seq(0,1,length.out=p)
window <- 1

# hypothesis testing
cursus(Y,Yloc,X,Xloc,window)

#> Analysing a single chromosome.

#> Error in checkForRemoteErrors(val): 2 nodes produced errors; first error: there is no package called 'globalSeq'