Internal function

This functions estimates the parameters of the negative binomial distribution by maximum likelihood. It is called by the functions omnibus and proprius.

intern.estim(y, offset = NULL)

Arguments

y
random variable: numeric vector of length n

offset
numeric vector of length n

Value

The function returns a list of numeric vectors.

Details

We assume the negative binomial distribution y_i ~ NB(mu,phi), where the samples are indexed by i (i=1,...,n). Our parametrisation leads to E[y]= mu and Var[y]= mu + phi*mu^2. With the an offset the model becomes y_i ~ NB(a_i*mu,phi), where the a_i are known.

References

A Rauschenberger, MA Jonker, MA van de Wiel, and RX Menezes (2016). "Testing for association between RNA-Seq and high-dimensional data", BMC Bioinformatics. 17:118. html pdf (open access)

See also

Examples

set.seed(1)
y <- rnbinom(n=1000,mu=10,size=1/0.2)
intern.estim(y)

#> $mu
#>    [1] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [11] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [21] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [31] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [41] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [51] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [61] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [71] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [81] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>   [91] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [101] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [111] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [121] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [131] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [141] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [151] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [161] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [171] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [181] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [191] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [201] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [211] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [221] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [231] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [241] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [251] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [261] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [271] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [281] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [291] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [301] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [311] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [321] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [331] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [341] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [351] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [361] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [371] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [381] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [391] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [401] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [411] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [421] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [431] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [441] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [451] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [461] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [471] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [481] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [491] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [501] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [511] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [521] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [531] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [541] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [551] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [561] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [571] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [581] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [591] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [601] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [611] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [621] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [631] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [641] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [651] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [661] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [671] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [681] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [691] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [701] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [711] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [721] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [731] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [741] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [751] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [761] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [771] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [781] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [791] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [801] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [811] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [821] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [831] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [841] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [851] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [861] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [871] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [881] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [891] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [901] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [911] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [921] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [931] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [941] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [951] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [961] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [971] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [981] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#>  [991] 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081 10.081
#> 
#> $phi
#> [1] 0.207415
#>